AI_Study_Planner/requirements.txt
streamlit
pandas
altair
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
pydantic[email]
cryptography
